requirements.txt
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
telebot==0.0.5

tensorflow
pillow
numpy
pandas
matplotlib
